+++ src/main/components/instr/VGMInstrSet.h
@@ -26,7 +26,7 @@ public:
   virtual bool GetInstrPointers();
   virtual bool LoadInstrs();
 
-  VGMInstr *AddInstr(uint32_t offset, uint32_t length, unsigned long bank, unsigned long instrNum,
+  VGMInstr *AddInstr(uint32_t offset, uint32_t length, uint32_t bank, uint32_t instrNum,
                      const std::string &instrName = "");
 
   std::vector<VGMInstr *> aInstrs;
