+++ src/main/formats/CPS/CPSTrackV2.h
@@ -57,7 +57,7 @@ enum CPSv2SeqEventType {
 class CPSTrackV2
     : public SeqTrack {
 public:
-  CPSTrackV2(CPSSeq *parentSeq, long offset = 0, long length = 0);
+  CPSTrackV2(CPSSeq *parentSeq, uint32_t offset = 0, uint32_t length = 0);
   void ResetVars() override;
   bool ReadEvent() override;
