+++ src/main/formats/HudsonSnes/HudsonSnesSeq.h
@@ -140,7 +140,7 @@ class HudsonSnesSeq
 class HudsonSnesTrack
     : public SeqTrack {
  public:
-  HudsonSnesTrack(HudsonSnesSeq *parentFile, long offset = 0, long length = 0);
+  HudsonSnesTrack(HudsonSnesSeq *parentFile, uint32_t offset = 0, uint32_t length = 0);
   void ResetVars() override;
   bool ReadEvent() override;
