+++ src/main/formats/MP2k/MP2kSeq.h
@@ -21,7 +21,7 @@ public:
 
 class MP2kTrack final : public SeqTrack {
 public:
-  MP2kTrack(MP2kSeq *parentSeq, long offset = 0, long length = 0);
+  MP2kTrack(MP2kSeq *parentSeq, uint32_t offset = 0, uint32_t length = 0);
 
   bool ReadEvent() override;
