+++ src/main/formats/NeverlandSnes/NeverlandSnesSeq.h
@@ -37,7 +37,7 @@ class NeverlandSnesSeq
 class NeverlandSnesTrack
     : public SeqTrack {
  public:
-  NeverlandSnesTrack(NeverlandSnesSeq *parentFile, long offset = 0, long length = 0);
+  NeverlandSnesTrack(NeverlandSnesSeq *parentFile, uint32_t offset = 0, uint32_t length = 0);
   virtual void ResetVars(void);
   virtual bool ReadEvent(void);
