+++ src/main/formats/SoftCreatSnes/SoftCreatSnesSeq.cpp
@@ -79,7 +79,7 @@ void SoftCreatSnesSeq::LoadEventMap() {
 //  SoftCreatSnesTrack
 //  ******************
 
-SoftCreatSnesTrack::SoftCreatSnesTrack(SoftCreatSnesSeq *parentFile, long offset, long length)
+SoftCreatSnesTrack::SoftCreatSnesTrack(SoftCreatSnesSeq *parentFile, uint32_t offset, uint32_t length)
     : SeqTrack(parentFile, offset, length) {
   ResetVars();
   bDetermineTrackLengthEventByEvent = true;
