+++ src/main/formats/SoftCreatSnes/SoftCreatSnesSeq.h
@@ -40,7 +40,7 @@ class SoftCreatSnesSeq
 class SoftCreatSnesTrack
     : public SeqTrack {
  public:
-  SoftCreatSnesTrack(SoftCreatSnesSeq *parentFile, long offset = 0, long length = 0);
+  SoftCreatSnesTrack(SoftCreatSnesSeq *parentFile, uint32_t offset = 0, uint32_t length = 0);
   virtual void ResetVars(void);
   virtual bool ReadEvent(void);
 };
