+++ src/main/formats/SuzukiSnes/SuzukiSnesSeq.h
@@ -98,7 +98,7 @@ class SuzukiSnesSeq
 class SuzukiSnesTrack
     : public SeqTrack {
  public:
-  SuzukiSnesTrack(SuzukiSnesSeq *parentFile, long offset = 0, long length = 0);
+  SuzukiSnesTrack(SuzukiSnesSeq *parentFile, uint32_t offset = 0, uint32_t length = 0);
   virtual void ResetVars(void);
   virtual bool ReadEvent(void);
