+++ src/main/formats/TriAcePS1/TriAcePS1Seq.h
@@ -41,7 +41,7 @@ class TriAcePS1ScorePattern
 class TriAcePS1Track
     : public SeqTrack {
  public:
-  TriAcePS1Track(TriAcePS1Seq *parentSeq, long offset = 0, long length = 0);
+  TriAcePS1Track(TriAcePS1Seq *parentSeq, uint32_t offset = 0, uint32_t length = 0);
 
   virtual void LoadTrackMainLoop(uint32_t stopOffset, int32_t stopTime);
   uint32_t ReadScorePattern(uint32_t offset);
